Created a helper function for category Get api
