+++ Helper/api/categoryHelper.js
@@ -1,4 +1,5 @@
 import { keys } from '../../lib/keys';
+import prisma from '../../lib/prisma';
 import { s3 } from '../../lib/s3Client';
 
 const base64Data = (image) => {
@@ -12,12 +13,9 @@ const type = (image) => {
   return image.split(';')[0].split('/')[1];
 };
 
-let contents = [];
-
 // create;
 export const Post = async (body) => {
   const { title, content, image } = body;
-  console.log('type', type(image));
 
   const params = {
     Bucket: keys.aws.s3Bucket,
@@ -45,3 +43,37 @@ export const Post = async (body) => {
 };
 
 // read;
+export const Get = async (res) => {
+  let contents = [];
+  contents.length = 0;
+
+  await prisma.category.findMany().then(async (r) => {
+    if (r.length >= contents.length) {
+      let promises = r.map(async (item) => {
+        const params = {
+          Bucket: keys.aws.s3Bucket,
+          Key: item.s3BucketKey,
+        };
+        await s3
+          .getObject(params)
+          .promise()
+          .then((re) => {
+            contents.push({
+              id: item.id,
+              title: item.title,
+              description: item.description,
+              image: re.Body.toString('base64'),
+            });
+          });
+      });
+
+      Promise.all(promises).then(() => {
+        res.json({
+          data: contents,
+          status: 200,
+          message: 'All data retrieved successfully',
+        });
+      });
+    }
+  });
+};
